requirements.txt
torch>=2.0
tokenizers
fastapi
uvicorn[standard]
websockets
pandas
pyyaml
tqdm
googletrans==4.0.0-rc1
nltk
